Modify nvapi project headers so that the main.h is common stuff, by moving the D3D9Namespace into the cpp itself.  Move the dxgi.h up into the d3d9namespace so that it can be used with no conflict in other headers.
Fix two compile errors missing the d3d9namespace for DXI_Swap_chain so that it gets the right variant. Still from the 8.1 headers.

Alter the project settings to use Xinput9_1_0.lib as the input library, not xinput.lib which is for 1.4 variant and won't work for Win7 use. Using old Xinput9.1 works on Win8 as well.  Add the _WIN32_WINNT=_WIN32_WINNT_WIN7 preprocessor directive to make the Xinput.h switch to 9.1 for compilation, as we are targeting the older Win7.

This makes the NVapi.dll project compile and create a dll that depends upon Xinput9_1_.dll.