Switch to unbuffered logging.
There were over 1000 fflush calls for logging, and with that much logging it makes more sense to use unbuffered I/O so that fflush is unnecessary.

Removed all fflush calls, and replaced them with blank lines. This serves two purposes. First, it removes the fflush itself, then second, it adds whitespace immediately after a log line which separates logging from functions. This makes it more readable by adding whitespace and reducing clutter.

The only possible drawback would be performance, as it can theoretically lower performance to remove buffering. If this is a problem, it would be better to add a fflush to the RunFrameActions for a 1/60 second timer.

This is a very large change affecting nearly every file. Done by global search and replace. Should be no actual impact on the code.