Add unbuffered flag in d3dx.ini for logging.
Add 'unbuffered' option in Logging section so that we can disable unbuffered if we want.  Not sure it matters, testing didn't show any drawback to unbuffered logging, but in a Release version it might be nice. Easy to add.

Added a single fflush(Logfile) in the RunFrameActions routine, so that at every frame the log is flushed as a compromise.