Add optional wait for debugger at wrapper launch.
It's a pain to attach after JIT with 2 screens in 3D, so this adds an optional wait loop at initialization until debugger is present.

It's triggered by a d3dx.ini setting that will not be in the default file of "waitfordebugger=1".  It will be a hidden feature for debugging.