re-add convenience copy for debugging.
Make a CopyToGames.bat file to hold xcopy commands to be executed upon PostBuild Event. This ensures that the we are always using the latest build with no manual steps necessary.

Commented out to start with, to avoid affecting others.

Put as many xcopy in here as you like.  Two examples in the .bat are for AC3 and Bioshock as the two primary test cases.