Fix Decompiler bug that would generate vector mismatch Asserts.
Occasionally, we'd get an Assert of vector mismatch, usually during a new game load in AC3.
The bug appears to be a memory corruption caused by the vector moving in memory and invalidating a stale pointer.
This fix breaks up the code so that the stale data cannot be stored.