Switch Release build to no longer use Debug runtime.
We were using Debug even for Release builds because of serious crashes.  Those are now fixed, and it seems to run smoothly without crashes using MultiThreaded, instead of MultiThreadedDebug.

Keeping it as static library still.  Does not seem to be much advantage to make it DLL..  Each subproject is bigger than it needs to be, but the dependcy on system dlls seem bad.