Add AC3 shaders as a Solution Item folder.
This first batch are the remaining shaders from the ShaderCache folder that do not recompile properly when running the game.  Including both the master ASM file generated, as well as the HLSL code generated by the Decompiler.  These will be useful to track down the remaining Decompiler bugs, as well as create a working set of shaders for ShaderFixes.

This allows us to put the shader work under source control, keep track of changes, share items easily, and do diffs upon shader changes.