Add new 'hunting' flag in d3dx.ini for performance.
hunting=1 when you want to find shaders and textures.
hunting=0 when you want the best performance.

This skips past all the saving and matching code on the shaders and textures, so that hunting is ignored as a Release mode.

In AC3, this bumped performance from min 27 fps to 33 fps.

It's not exactly right, as it disables some features like Texture remapping from the .ini.  We don't use that in AC3, but it is used in BI.  OK for now.