Add an experimental thread based hunting loop.
This is a thread that sleeps for 100ms, then runs the RunFrameActions loop.

It's experimental because I'm not certain how it interacts with other parts of the tool.  I expect it to be OK, where it collects 100ms of shaders, instead of 15ms for one frame, so somewhat more than usual, but still a small subset of what is live.

Seems to allow hunting and marking with no problems.