Improve thread based hunting by slowing buffer clearing.
The prior technique was to clear the hunting buffer every VBL, which is a bit too often, because things animate and move in a given scene.

This change makes the clear operation only happen after 1 minute of no user input for hunting.  Thus, the hunting buffers for pixel and vertex shaders stay the same and the hunting is much more stable.