Fix reversed opcodes for ishr, ishl, ushr.
The documentation online for ishr/ishl is backwards, and so this code used that and would generate shift operations that were backwards.

This reverses op2 and op3 in the output so that it generates the C style code for shift operations in HLSL, which FXC compiles to the right code, as tested by doing direct ASM comparisons of master code to HLSL compiled code.