Fix Decompiler bug for float4 matrices as arrays.
Format was row_major float4x4 name[N];

The code did not expect those ZxY matrices to ever be arrays, and thus was putting in [6] instead of [0] for a 6 element array.
This change adds them to the name list array of mCBufferData, so that every possible numeric element is named properly.
Tested on AC4 code, and works as desired.

Deleted the sample shaders in AC4/ as being fixed.
But leaving the hand fixed ones in AC4/ShaderFixes, to avoid changing anything regarding the actual fix for the game.  Code should be the same anyway.