Move uninstall.bat to dependencies, and complete it.
Didn't uninstall all the possible debug stuff, and so it now includes every possible thing that winds up in the game folder.
Also deletes the ShaderFixes folder.

Moved to dependencies so that it gets copied for local builds.

Move specific copy step for game target builds, as it gets copied with dependencies.