Temporary changes to work with FEAR3.
Crash at start with d3d9, so current step is to disable building d3d9 for the project.  Not sure it's necessary in any case, and being missing does not seem to cause problems.

Switch back to thread based shader hunting, as dxgi version does not work (possibly from lack of d3d9.)