Simplify project releases to just 3 DLLs.
This dramatically simplifies the release, by removing and deprecating older project pieces that do not seem to be neeeded.

Remove all D3DCompiler_xx projects, as these create a wrapped compiler which is not used for anything.
Remove D3D10 and D3D9 projects as these wrap both those DLLs, but for DX11 games it does not appear to be necessary, and just adds confusion.
Remove DXGI project as it causes problems on Win8.1, and forces us to use the KB update to fix crashes.  Also, it does not seem to be used for anything other than getting acess to the Present call, which is now handled in Context->Draw.

This requires a new xcopy of the Win8 redist DLL for the D3DCompiler_46.dll, the only one we seem to use or need.

Tested in AC4, SR4, and it still works as expected.