Complete first stage of SR3 fix, for texture alignment.
This fixes all the VS that have texture alignment problems in stereo, using the auto-fix of 3Dmigoto.

This might introduce some bugs, as some of the texture outputs don't seem to necessarily use o0 as a match.  But this is a good step in terms of finding every shader affected, as they are auto-fixed upon being found.

There may be more later in the game, this is from New Game to first Crib.