Fix a bunch of x32 warnings.
Doing x32 warnings first, as they'll be in x64 as well.

This fixes a bunch of sscanf_s warnings, all of them that don't need %s or %c are just changed as nothing needs to be added.

Fix some sprintf_s where a size is needed, by making some global sizes that we can use for things like opcodes and strings, where they are passed as char* params, and thus not available otherwise.

Use size_t for everything that might be signed/unsigned problems, as that will work in x64.