Fix DirectInput use of NVAPI project.
It was using the wrong version of DirectInput.h, from the project root, which made the aiming override fail because it cannot see proper keypresses.

This changes the project to use the proper file.

While this is using the right DirectInput, we are also adding a new thread based timer here, so that the nvapi can get time slices independent of the d3d11, so that it can set convergence as necessary.
The old code was doing this on every SetConvergence call, but in Ghosts, that call is not made on a regular basis, and it's not a good idea to rely upon that.