Fix nvapi.lib dependencies, make it load time dll.
We need the root nvapi.lib to link, so those are re-added to the project files.

Also create a call to nvapi_queryinterface so that it requires the dll at launch of the game.  That will preload the proper nvapi.dll from our directory, and force it to the front of the line.

This makes it a launch dependency, which is what we want.