Set up to remove DebugLayer for CreateDevice. Amended last commit to add this file.
In Debug we usually set D3D11_CREATE_dEVICE_DEBUG at CreateDevice, so that we get the debug layer.  For end-users though, they probably do not have the SDK installed, and this errors out.

Temporarly set this to off, so we can let people use a Debug build to find a crash with Ghosts.