Update the MainHook to log better, and use LoadLibraryExW.
The APImonitor showed it was LoadLibraryExW, so switched to that variant.
Restructured it so that DLLMain is setup with the normal switch statement.

Add regular d3d11_log.txt logging for anything actually hooked, so we can see if it's working and being called.

With a proper matching case, this now works in WatchDogs!