Set up WatchDogs target to be only one that hooks.
This only allows the deviare hooking and bypassing the system32 lookup in the WATCH_DOGS target only.  Maybe later we extend it, but it's not necessary anywhere else yet.

Fix a bug with looking for "original_d3d11.dll" as the target specified from wrapped d3d11 so that it can get real version without calling back into itself.

Make the WatchDogs build target create a real release, including dependencies.