Fix two fatal errors in WatchDogs Decompile.
First was missing RetC opcode, for conditional return.  Parse failure. Code is not identical, but should work the same.
Second was Texture2DMS with a 0 length, causing it to generate random numbers.  Not legal syntax according to the docs, but setting it to just Texture2DMS<float4> makes it generate the same code as ASM.