BinaryDecompiler: Backport "Further work on hull shader phases"
Previous code was limited to a single join phase and no more than two
fork phases. Support for input and output patch constants is still
incomplete.

DarkStarSword notes:
For the backport, changed [p]psDecl/[p]psInstr to be vectors to match
the existing 3DMigoto port, removed separate count variables. Used
resize() for some allocations and push_back to append to vectors.

This resolves the crash in Splinter Cell Blacklist and the example hull
shader no longer throws an exception (the output shader is missing the
body, but that's to be expected since the HLSL decompiler does not yet
support hull shaders).