Fix a slowdown that 3Dmigoto introduces to games.
This was missed during the earlier ShaderHunting search for code to excise during run-mode.  The CreateBuffer is wrapped and was doing a hash calculation on the buffer, which was taking some 3% of the CPU.

This change just removes that calculation when shaderhunting=0, and makes CPU usage at 0.8%.
Also fixes a stall in AC4 that would cause terrible Min frame rates.