Fix boolean checks used for AND opcodes.
The prior version would look for only register use, like r1, not including components like r1.xyw.  That failed in an Alien Isolation case, because the r1.x use cleared the boolean before the r1.y was used.

This makes it use the full register+components as a single string for comparison.

Also changes all the unrolled instructions to be rolled back up, as those unrolled ones were damaging the original op1 string.

Also fixes an error in BFI instruction that wasn't incrementing the constants.