Update build target for AC4 to latest code.
The old version used the 7 dll approach, this now uses the 3 dll approach.
Clean up the build target to use only Win8 libraries and headers.
Compiles and builds now with no warnings.

The reason to update this is to include the new fix for AC4 which fixes the performance glitch walking past the bottles.