Remove the export_binary option from d3dx.ini.
This would generate .bin files for every shader seen, and is something that doesn't seem to make sense. We never used this feature, and I'm unsure when I would ever use this, or why.
Especially because there is already a cache_shaders option which does that for active shaders, but not every one seen.

This change is part of making ShaderCache just an archive, and not live.