Fix the feature of aiming override changing separation.
The code that was there did not seem to function properly, so this rewrites that to use CPU time from the Draw call hook.  That seems to be the most stable time to look for mouse button held down, and calling SetSeparation or SetConvergence.

Probably will restructure this later, but works for now.