Clean up some excessive logging, add Touch operation to CopyOnMark.
Some excessive logging was moved to Debug log instead, especially the QueryInterface calls.  Goal is to only log errors, not everything, unless we are in debug log.

Add Touch operation to the CopyOnMark for files that already exist, instead of making it an error.  File is there, it's good, we don't overwrite what is there. Touching the file makes it sort to the top of most recent.