Add dynamic iniParams to wrapper.
This does the work to Map the iniParams back to CPU, rewrite them to zero, then UnMap them back to GPU.

Works to set the iniParams to zero, from whatever they are read as.

Code tested, and works, including the toggle. But commented out for now, as it's hard coded to Esc key.

Also comment out the experimental aim override via nvapi.dll, as questionable.