Make HOOK_SYSTEM32=1 the default for x64 builds.
This changes the default state for Release builds to enable the hooking.
Has not seemed to be a problem so far, and having it not enabled has caused confusion for end-users.
Changing default to always on, until we see if anything bad happens.

Off by default for x32 builds and game specific builds need to specify it when needed.