Change input processing to use a callback model
The actions list is no longer statically allocated - it's now a vector
of struct Actions, which allows us to dynamically add new bindings on
the fly, which will be useful to parse arbitrary custom keybindings from
the ini.

Inputs are now registered at run-time via register_ini_key_binding(),
which takes care of reading the binding from the ini file and adding it
to the actions list.