Fix iterator dereference past end of list
I should have known that combining a C style for each macro with C++
vectors would trip me up - the problem here is the iterator dereference
in the increment part of the loop is run before the condition, so will
dereference the iterator past the end of the vector.

I've moved the dereference into the condition part, using the &&
operator so it is only run if the condition is true. Since assignments
return the value of the assignment (or am I making C assumptions
again?), the whole condition still works.

But maybe I should just kill this macro and embrace C++ instead...?