Fix convertHexToFloat to handle integer literals.
Turns out we can have l(-1) for an AND statement as well, so that would then convert badly into float.

This fixes that problem by making a second path in the case where the literals are not hex, that is, do not start with 0x.

Fixes a QNAN problem in FC4.