Far Cry 4: Small improvement to nVidia god rays (not accurate, 1920x1080 only)
This simply pushes them back to infinity, but for an accurate fix we
need to calculate them correctly in each eye to begin with.

In order to remove the halo that resulted I adjusted a 2nd related
shader, but it's scale was in pixels so I've hardcoded the resolution as
1920x1080 until the screen resolution functionality in 3Dmigoto is
restored.

This is pretty low priority compared to everything else though.