Refactor logging & recording render target info
Setting shader resources and the render target were all doing virtually
the same thing - the same thing I want to do yet again when drawing, so
refactor them all to use common functions.

As part of this, also change the struct ResourceInfo to use the
assignment operator, which allows it's contents to be filled out later
and works better with the new code flow.