Include SampleDesc in Texture2D hash (NOTE: CHANGES TEXTURE/RENDER TARGET HASHES!)
The SampleDesc info was missing from these hashes, which could make them
less unique than they should be. I also want this to be able to use the
hash as the map key when storing their descriptions to avoid storing
multiple copies of the same texture description, which needs a 1:1
relationship between the hash and the texture descriptions to be safe.

This will change the hashes of any render targets, which is not really
desirable, but it looks like none of the fixes in the 3Dmigoto
repository are using it at the moment, so it's probably better to change
it now rather than later.

If the worst comes to the worst and this breaks someone's fix, we can
always add a flag in the ini file to select the old hash instead.