Refactor CreateTextureND to use the same hash functions (NOTE: Changed 3D texture hash!)
These hash functions are almost identical with only slight differences:
- The create texture versions can use the texture's initial data, which
  we can't do in the others (maybe we could use the current data though)
- The create texture versions use special values for the width and
  height if they match the screen resolution, which the other's don't.
  It seems like a good idea - perhaps we should change the others to do
  the same thing?
- The CreateTexture3D used the height twice instead of the depth - this
  is clearly a bug.

Refactor them to use the same code path so we don't get bitten if we
change the hash in only one place later, but note my concerns about the
existing slight differences in the function.

Refactoring the Texture3D hash will inevitably change it, as the
previous code has a bug which uses the height twice in the hash instead
of the depth. This is unfortunate, but I don't know of anyone using this
currently and it's clearly a bug, so I'd rather fix it than leave it
broken. If this does break someone's fix we can always add a flag to
revert to the old hash behaviour.