Use hash as key for render target descriptions, rather than handle
The hash includes every field in the texture description, so assuming
the hash function is good there should be a 1 to 1 relationship between
hashes and unique texture descriptions.

Using it as the key rather than the handle means we won't store
additional descriptions for textures with identical properties, and if a
game were to keep creating and destroying render targets at different
memory locations we won't have to worry about filling up memory with
more and more identical texture descriptions.