Collect shader resource usage stats on draw calls
The ShaderUsage.txt seemed to be largely incomplete - namely, it was
missing most of the texture resources that I've been trying to identify.

The issue I believe was because 3Dmigoto was only associating resources
with the active shaders when they were set, but they can also be used by
other shaders set afterwards unless they are explicitly cleared.

By collecting resource usage stats on the draw calls, we can be sure to
get all the resources that are still bound to the pipeline and may be
used by the active shaders, no matter how long ago they were set.