Always record shader resources, even if not tex2d/3d
Other resources types may be of interest as well, even if we haven't
collected specific information about them. We will need to add some
minimal information to the ShaderUsage.txt to be useful (as at the
moment their hash is 0 which won't match up to anything else), but this
will at least signify that there is *a* resource in a particular slot.

I'm still not seeing all the resources I expected, so I need to keep
looking.