Move log wrappers to generic file for use in other source files
I have not made any attempt to clean up the LogFile / LogDebug
declarations, and it doesn't seem like it would gain much unless it
happens as part of a move to a full logging framework.

This means that these will continue to be used however they already are
in the source file this is included from, be that either in the local
file, an extern or accessing another namespace.