Don't clobber recorded render targets for a shader
This allows us to see if a shader has been used with different render
targets at different times. At the moment it stores the render targets
as a set of vectors. That works, but can make the ShaderUsage.txt overly
verbose if (say) a shader had been used with 5 render targets
simultaneously, then another 5, only four of which matched the first
set.

Reversing the way this is stored should make things clearer, but I
wanted to commit this first since it is working.