Add missing sample_aoffimmi handler. Switch IEQ.
No sample_aoffimmi handler before, and it failed in FC4 by making a green cast to the scene. This seems to fix the problem there, and generates the same ASM for the sample instructions.

Also changed IEQ and INE to be different than EQ, both fell into the same switch case.  IEQ/INE now converts to int, so that it doesn't do float compares for the int case.  Not the problem here, but is a potential problem to miss compares that are off by small amounts.