Fix named array references in Decompiler. Risky.
This is a big change in the way that applySwizzle works, as I reworked the code for handling constant buffer named variables.  There are numerous variants, and this seems like a more solid way to handle the variants.

The problem was cb2[r13.y + 44].x type instructions were being truncated to cb2[r13.y.x as errors.  To fix that, this handles the naming more correctly, but dramatically changes other pieces as well.

This also fixes an different problem seen when there is no name for variable, it gets set as cb0[42] variants.  If those are found, we let the code through as is, because we have no better name.

Also fixes a sign problem, where it would drop the negative sign for constant buffer items only.  Unified that to be across all variables input.  Risky, but otherwise is too complicated and fragile.

Compiles and runs in FC4, fixes the problems seen in shaders there.