Fix bugs in resinfo opcode parsing.
Numerous problems with the previous version, this rewrites it completely using the BinaryCompiler flags as much as possible.
This uses temporary registers defined at the top, to hold the return values, because this instruction GetDimension does not have a return value, only parameters.
Those parameters are then assigned to what the output was defined to be.

Should handle the swizzle properly. Not 100% sure, it's fairly gnarly.
Definitely handles both float and uint variants, and should generate a non-compiling error for cases that don't work, instead of generating bad code.

Also added some code to the BinaryCompiler, copied from the newest version on Git, to handle the resinfo flags that were not previously decoded.