Fix 'udiv' opcode handler for constants.
Was doing double applySwizzle on the opcodes, which doesn't work for constant values.
So, now copy the opcodes and make a second copy of the src0 and src1 assigns to be certain that the swizzle applies.
Documentation is unclear on whether swizzle must be the same or can vary between halves, so this is the safe approach.

Also fix a problem with a buffer in applySwizzle that was too small to handle input parameters with array outputs.
Set all character arrays to the 128 character opcodeSize, for consistency.