One more try on the 'udiv'.  Simplification.
This change removes the use of src0 and src1, and instead puts the first value into uiDest, then assigns the output at the tail.

This also fixes a problem where it was getting errors converting float2 to uint4.

The cast operator of (uintN) is used to force the instruction to recognize the 'udiv' and not just be a regular divide.

Tested on FC4 and seems to handle all cases there, including the constant l(6) as divisor, and swizzling of inputs.