Add special sampler_c handler for empty Resource Bindings.
In the case where debug/name info is stripped from a shader, we still see the dcl_sample section and add in default names likes s0_s.

This is done for both mode_default and mode_comparison, to create the correct naming type.

This fixes all the SampleCmpLevelZero errors in FC4.