Replace DirectInput with new GetAsyncKeyState input infrastructure
The old DirectInput backend was problematic - for some games it didn't
work at all, and it used the names of the keys in the current users
language, making it impossible to ship an ini file and expect the
keybindings to work on everyone's system.

This patch completely removes it from the DirectX11 project, and
replaces it with a fairly simple infrastructure based around the
GetAsyncKeyState() call.

Since this lives entirely within the DirectX11 project, the void* used
for the device has been returned to using the proper type - if we later
wanted to share this code with another project we should find a better
way to abstract that.

Keys can be specified in a number of ways - the letters A-Z and 0-9 on
the number row are simply specified as a single case independent
character, e.g.

next_pixelshader=k
previous_pixelshader=J
mark_pixelshader=9

Other keys are specified using the virtual key name as found in this
MSDN article, case independent with or without the VK_ prefix:
http://msdn.microsoft.com/en-us/library/windows/desktop/dd375731(v=vs.85).aspx

e.g.
next_vertexshader=VK_NUMPAD5
previous_vertexshader=NUMPAD4
mark_vertexshader=Numpad6

The raw hex codes can also be used with a 0x prefix, e.g.
mark_vertexshader=0x28

Some aliases are provided for the default keybindings English names as
used by DirectInput, so users with an old d3dx.ini don't have to change
anything (provided they are only using the default English keybindings).

The bindings in the default d3dx.ini have been updated with the new
names for each of the bindings with a link to the MSDN article (as
above, the old names would continue to work, but this should make it
more obvious to anyone looking to customise them). The examples for
using a controller have been removed since they will no longer work.

If we wanted to add controller support back in we could consider adding
another backend to the new infrastructure (either the previous
DirectInput support, which we would naturally disable by default, or
perhaps something new based on xinput). The new infrastructure includes
an abstraction in the code where I think it would make sense to separate
out the generic infrastructure and a backend specific part of an input
action, but the backends themselves would also need to be abstracted.