Support for convergence & separation presets loaded from the d3dx.ini
Convergence & separation presets can now be configured in the d3dx.ini
file on arbitrary keyboard and mouse buttons, for example:

[Key1]
Key=RBUTTON
convergence=0.1
type=hold

[Key2]
Key=q
separation=0.1
type=toggle

[Key3]
Key=z
separation=100.0
convergence=4.0

Three types are supported - by default the bindings will simply load the
configured settings, but type=hold can be specified to have a preset
active while the button is held, and type=toggle can be used to make a
simple on/off toggle. Later we may also add an ability to cycle between
several presets from a single button.

For now this still uses the input callback infrastructure, but this is
not necessarily final. Due to a limitation in C++ it was necessary to
use static proxy methods for the callbacks to call the instance methods
- these are fairly concise, so we can probably leave it as is, but I
generally dislike unnecessary boilerplate code, and am considering an
alternate design that would avoid the need for them.

This patch also includes a few smaller changes as part of this:
- NvAPI_Stereo_SetConvergence() checks gDirectXOverride - this seemed to
  be necessary for the convergence presets to work.
- RegisterIniKeyBinding() has been split into two functions to avoid
  having to read the Key= field from each key section twice.
- log_file is no longer passed to the RegisterIniKeyBinding() - this was
  never necessary, and only existed as I had not realised that LogFile
  was global at the time.