Replace input callback mechanism with an InputListener interface
This defines a new interface called InputListener, which can receive key
up & down events. A new KeyOverride class uses multiple inheritance to
implement an Override that can receive key events.

When an InputListener is registered, it is combind with an InputAction
that handles the details of checking the input device state, and calls
the DownEvent & UpEvent functions on the listener when appropriate.

For simplicity with key bindings that don't otherwise need their own
object (e.g. most of the shader hunting keys), a generic InputCallbacks
class implements the InputListener interface and dispatches events via a
simple callback. RegisterIniKeyBinding hides this implementation detail
so that the existing bindings continue to work unchanged.