Move globals and related declarations into globals.h
At the moment these globals are static and only available within a
single compilation unit and #include is abused to make these seem to be
available in other files as well.

This is the first step in allowing these to be accessed from other files
in the more usual way, by moving the declaration of the Globals structure
(and those it uses) out into a separate header and exporting the globals
externally.

In particular this is necessary to access the global iniParams from
Override.cpp in the next commit.