Create a full Zip Release to publish final builds.
This fully automates the build for creating final builds to be released.

The Zip Release target builds the x64 or Win32 variants and creates a valid output folder.

The new Publish.bat batch file will clean the outputs completely, rebuild both Win32 and x64, then move the results to the zip output directory and zip up a final release.