Fix bug with iniParam handling that required .x first.
Now check for all 4 parameters so they are independent of each other and does not require .x to be used first.

Set iniParams to be FLT_MAX as the unused state instead of -1, to match Override usage.

Add better logging to Override operations for on/off states, with better output formatting.