Add unlock_separation setting in d3dx.ini
This is useful for CryEngine games that insist on resetting the
separation to 50% every frame to allow users to set their preferred
separation value instead.

Convergence unlocking is not added as it appears the existing
convergence map code may already do this and I don't want to touch that
without something to test it with.