Add experimental support for reloading the d3dx.ini live
When the reload_config key is pressed, the entire block of code
responsible for loading the d3dx.ini is run again. That code was not
originally designed with reloading in mind, so certain settings
will not be cleared if they have been disabled or commented out in the
ini. I will address these issues in a separate patch for a clearer diff.

Key bindings are cleared, which is not safe to do while iterating over
them in the input subsystem, so the config reload is delayed until after
other keys have been processed.

This support is disabled by default for now until it has got some more
testing.