Clear Overrides and Constants when reloading config
This means that if a user removes a ShaderOverride, TextureOverride or
Constant setting and reloads the config file the setting should get
cleared as they expect.

There are some other settings that may also benefit from being
explicitly cleared like this, but it's not clear if those settings even
make sense to reload in the first place, as many of them will only
affect the initialisation path or WritePatches() routine anyway.