Note an example of a setting that only partially works when changed live
There's no doubt going to be more examples like this, where a setting
must have been enabled during initialisation to work properly so enabling
it later either won't work or will only partially work.

e.g. hunting can be enabled live, allowing shaders to be cycled with the
key bindings and the hashes listed in the log file, but any shaders that
have already been created by the game won't be registered for reload and
therefore will fail to copy to the ShaderFixes directory when marked.

e.g. some of the marking modes must be set when the shaders are created,
so changing that setting live will not work as expected.

None of these seem fatal though, and it seems that there is a lot that
works just fine if reloaded live, so I don't think these are show
stoppers, just caveats to keep in mind. We could improve some of them,
but any that might trade off performance or memory usage in production
mode are probably not worth while.