Add VKRepeatingInputAction class for auto-repeat.
This implements auto-repeat for the hunting keys only, by making a new subclass of VKInputAction that supports auto-repeat in the Dispatch method override.

If repeat_rate is set to nonzero, it will create a VKRepeatingInputAction instead of a VKInputAction object.

repeat_rate can be changed in the d3dx.ini file in in the Hunting section to change the speed.

As a first pass this seems to work OK, but might need tuning as we get experience with it.