Clarify project licensing due to inclusion of GPLv3 Deviare library
As 3Dmigoto makes use of the Deviare-InProcess library licensed under
the GPLv3, a concern has been raised that 3Dmigoto must also be licensed
under the GPLv3 and may have consequences if someone were to include
3Dmigoto in a commercial closed source product. I believe that this is
true in certain (but not all) situations, so I have updated the
licensing and copying information to better reflect this situation.

In particular, whenever 3Dmigoto is distributed in a form where it is
linked against the Deviare library, it will be covered by the terms of
the GPLv3 - with all the requirements thereof (such as providing source
code of 3Dmigoto AND the Deviare library to anyone who requests it
within 3 years of distribution).

The source code of 3Dmigoto however, is still validly licensed under the
more permissive MIT license, and as such someone may be able to use it
without being bound by the terms of the GPLv3 by purchasing a license to
Deviare, or removing Deviare from 3Dmigoto.