First pass adding support for linear transitions on overrides
This extends the Override class, so all overrides may include an
optional transition, allowing for smoother changes over a period of time
rather than changing instantly.

This can be used in conjunction with a delay - the transition will occur
after the delay.

Each parameter can only have one active transition, so activating a new
transition will stop any transitions for the same parameters. This is
done to reduce edge cases.

At the moment transitions have a higher chance of messing up the value
to be restored. e.g. if a button is held then released, then held again
during the releasetransition period, the current partial transition
value will be saved as the value to restore on the next release. This is
another case of the more general problem of the restore values being
associated with a specific override instance, and may be better resolved
by using a global refcounted restore value for each parameter.