Add a global parameter save area + ignore intermediate transition values
This global save area does not replace the local saves in each override,
rather it is indented to ensure that the final restore for a given
parameter will restore it back to the original value if several
type=hold/toggle keys affecting the same parameter have been used and
released in a different order.

Additionally, if a parameter that is currently being used in a
transition is saved, the final value of the transition will be saved
instead of the current intermediate value. This resolves issues where it
was possible to e.g. permanently change the convergence by rapidly
pressing right mouse if it had a convergence override and used a
releasetransition=N.

The only situation I have identified where this is still sub-optimal is
when a type=toggle key is activated, then several type=hold keys
affecting the same parameter are held and released in a different order.
The original value will still be restored once all keys are deactivated
including the type=toggle, but until then the value with only the
type=toggle key active can become the value of one of the type=hold
keys. Potentially an undo stack replacing the local saves might help
here, but I'm not convinced that would be the right answer for other
situations and this seems like a rare enough edge case that it is
unlikely to matter in practice.