Correct the .RC files for DirectX11 and NVAPI projects for real versions.
This updates the .RC file for DirectX11, and adds a new RC file for NVAPI to use the correct version from the updated version.h file.

This will now apply a version number to both DLLs, that matches the project version and is viewable via properties.

This removes the wrong and misleading microsoft version and file info, and replaces it with correct info.   Built into the dll files now, yeah!

Tested with Mordor to be sure DLLs still load, and they do.

To use this, all we need to do is update version.h to new numbers.  Next step will be to automate the Publish to auto-increment.